refactor: update useMyPools hook
